Removed player pauses and "active" package selection
Removed the "in-between pause" feature from the Player. It was messy and overcomplicated. The components that use this feature will be refactored to simply add content blocks to the queue on the fly.

Also removed the GraphicsManager "active package" limitation. All graphics packages in the package directory will be imported at startup, and their layers can be referenced through a /Package/Layer tree structure.