Another gargantuan system overhaul!! (Player V2)
Hopefully the last one for a while 😬🤞
The player (and surrounding playback systems) has been almost entirely re-implemented to allow for real multi-track playback. This turned out to be much trickier than just "take two players and stack 'em up". Instead of a using a list of upcoming content like a sane person, the playback queue is now represented as a tree. Each level of the tree represents a different "layer" of content on screen, and content "nodes" can be attached to each other sequentially (like the good ol' fashion queues we know and stan) or attach somewhere in the middle (like "start playing me 30 seconds into this video"). 

If I did it right, the new tree system should allow me to properly implement all the cool features that were hacked in to the OG player. Playback events (which I want to rename to Rules) will be added back in next!